deprecate legacy async client and replace with async native
